+++ brain.lisp
@@ -37,10 +37,15 @@
   (setf *active-channel* (traverse-a-list message :from :channel))
   nil)
 
+(defun handle-event-gone (message)
+  (if (string= *active-channel* (traverse-a-list message :from :channel))
+      (setf *active-channel* nil)))
+
 (defun handle-event (message)
   (switch ((assoc-value :event message) :test 'string=)
           ("new-day" (handle-event-new-day message))
-          ("presence" (handle-event-presence message))))
+          ("presence" (handle-event-presence message))
+          ("gone" (handle-event-gone message))))
 
 (defun handle-user-message (message)
   (let* ((from (assoc-value :from message))