+++ test.lisp
@@ -13,12 +13,12 @@
 (setf prove:*enable-colors* nil)
 
 (if (and
+     (prove:run #P"tests/active_channel.lisp" :reporter :list)
      (prove:run #P"tests/active_user.lisp" :reporter :list)
      (prove:run #P"tests/commands.lisp" :reporter :list)
      (prove:run #P"tests/dont_understand.lisp" :reporter :list)
      (prove:run #P"tests/good_morning.lisp" :reporter :list)
      (prove:run #P"tests/handle_message.lisp" :reporter :list)
-     (prove:run #P"tests/hello.lisp" :reporter :list)
-     (prove:run #P"tests/presence.lisp" :reporter :list))
+     (prove:run #P"tests/hello.lisp" :reporter :list))
     (format t "PASSED~%")
   (format t "FAILED~%"))