+++ translate.lisp
@@ -7,9 +7,7 @@
   (setf *translator-io* (create-socket-stream socket-path)))
 
 (defun translate (message key result-key)
-  (json:encode-json-plist
-   (list key message)
-   *translator-io*)
+  (json:encode-json-plist (list key message) *translator-io*)
   (format *translator-io* "~%")
   (cdr (assoc result-key (json:decode-json *translator-io*))))